cleanup and reduce copies
