+++ Konscious.Security.Cryptography.Argon2/Argon2Lane.cs
@@ -1,7 +1,6 @@
 namespace Konscious.Security.Cryptography
 {
     using System;
-    using System.Runtime.InteropServices;
 
     internal class Argon2Lane
     {
@@ -27,38 +26,5 @@ namespace Konscious.Security.Cryptography
         public int BlockCount { get; }
 
         private readonly Memory<ulong> _memory;
-    }
-
-    internal static class SpanExtensions
-    {
-        public static void Blit(this Span<ulong> toBlit, ReadOnlySpan<byte> bytes, int destOffset = 0, int srcOffset = 0, int byteLength = -1)
-        {
-            int remainder = 0;
-            int length;
-            if (byteLength < 0)
-            {
-                length = 128;
-            }
-            else
-            {
-                length = byteLength / 8;
-                remainder = byteLength - (length * 8);
-            }
-            //---------------
-            var newSpan = MemoryMarshal.Cast<byte, ulong>(bytes[srcOffset..]);
-            newSpan.CopyTo(toBlit[destOffset..]);
-
-            if (remainder != 0)
-            {
-                var remainderSpan = bytes[^(byteLength % 8)..];
-
-                ulong extra = 0;
-                for (int i = 0; i < remainderSpan.Length; i++)
-                {
-                    extra |= ((ulong)remainderSpan[i]) << (8 * i);
-                }
-                toBlit[newSpan.Length + destOffset] = extra;
-            }
-        }
     }
 }
